requirements.txt
django==2.1.*
beautifulsoup4==4.8.2
lxml
pillow==7.1.0
pyppeteer==0.0.25
loguru==0.4.1
djangorestframework==3.9.1
requests==2.24.0
setuptools-scm
whoosh==2.7.4
django-haystack==3.0
Markdown==3.3.3
jieba==0.42.1
mammoth==1.4.13
markdownify==0.6.0
pyexecjs==1.5.1
psutil==5.8.0
PyQt5==5.11.3
PyYAML==5.4.1
arrow==0.17.0
pymysql==1.0.2
qiniu==7.3.1
Markdown==3.3.3
selenium==3.141.0
webdriver_manager==3.3.0
PyYAML==5.4.1
